Fix boxplot example to render data correctly
Original author incorrectly assumed that the e-charts library would compute the
quartile values for you. However according to
https://echarts.apache.org/en/option.html#series-boxplot.data and subsequent
verification with the example data I determined it was not doing this. Thus this
patch now exists to show a correct working example of box plots with the original
example provided by the author.